+++ CHANGELOG.md
@@ -35,7 +35,8 @@ Those two types were included in Nuke to make integrating third party caching li
 - `Loader` constructor now provides a default value for `DataDecoding` object
 - `DataLoading` protocol now works with a `Nuke.Request` and not `URLRequest` in case some extra info from `URLRequest` is required
 - Reduce default `URLCache` disk capacity from 200 MB to 150 MB
-- Reduce default `maxConcurrentOperationCount` of `DataLoader` from 8 to 6.
+- Reduce default `maxConcurrentOperationCount` of `DataLoader` from 8 to 6
+- Shared objects (like `Manager.shared`) are now constants.
 - Improved documentation